+++ src/app/chat/(mobile)/mobile/layout.mobile.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { PropsWithChildren } from 'react';
+
+import AppLayoutMobile from '@/layout/AppLayout.mobile';
+
+import ChatHeader from '../features/ChatHeader';
+
+export default ({ children }: PropsWithChildren) => (
+  <AppLayoutMobile navBar={<ChatHeader />}>{children}</AppLayoutMobile>
+);