+++ src/app/settings/(mobile)/index.tsx
@@ -1,0 +1,3 @@
+import Index from './mobile';
+
+export default () => <Index />;