+++ src/locales/options.ts
@@ -1,4 +1,3 @@
-
 import type { Locales } from '@/types/locale';
 
 type LocaleOptions = {
@@ -18,6 +17,14 @@ export const localeOptions: LocaleOptions = [
   {
     label: 'English',
     value: 'en-US',
+  },
+  {
+    label: '日本語',
+    value: 'ja-JP',
+  },
+  {
+    label: '한국어',
+    value: 'ko-KR',
   },
   {
     label: 'Russian',