+++ src/locales/resources/ja_JP.ts
@@ -1,0 +1,19 @@
+import common from '../../../locales/ja_JP/common.json';
+import empty from '../../../locales/ja_JP/empty.json';
+import error from '../../../locales/ja_JP/error.json';
+import market from '../../../locales/ja_JP/market.json';
+import plugin from '../../../locales/ja_JP/plugin.json';
+import setting from '../../../locales/ja_JP/setting.json';
+import welcome from '../../../locales/ja_JP/welcome.json';
+
+const resources = {
+  common,
+  empty,
+  error,
+  market,
+  plugin,
+  setting,
+  welcome,
+} as const;
+
+export default resources;