+++ src/locales/resources/ko_KR.ts
@@ -1,0 +1,19 @@
+import common from '../../../locales/ko_KR/common.json';
+import empty from '../../../locales/ko_KR/empty.json';
+import error from '../../../locales/ko_KR/error.json';
+import market from '../../../locales/ko_KR/market.json';
+import plugin from '../../../locales/ko_KR/plugin.json';
+import setting from '../../../locales/ko_KR/setting.json';
+import welcome from '../../../locales/ko_KR/welcome.json';
+
+const resources = {
+  common,
+  empty,
+  error,
+  market,
+  plugin,
+  setting,
+  welcome,
+} as const;
+
+export default resources;